# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LedgerGuard API",
    "version": "1.0.0",
    "description": "Read-only payment integrity control plane. Browser mutations require a CSRF token and idempotency key; versioned mutations also require If-Match."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "Session": []
    }
  ],
  "components": {
    "securitySchemes": {
      "Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sessionid"
      },
      "WooSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-LedgerGuard-Signature"
      },
      "StripeSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Stripe-Signature"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 0
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "correlation_id": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "correlation_id"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/api/v1/stores/{store_id}/pairing-codes": {
      "post": {
        "operationId": "pairing_code",
        "summary": "Pairing code",
        "parameters": [
          {
            "name": "store_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRFToken",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pairing_code": {
                      "type": "string"
                    },
                    "expires_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "mode": {
                      "enum": [
                        "test",
                        "live"
                      ]
                    }
                  },
                  "required": [
                    "pairing_code",
                    "expires_at",
                    "mode"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; no source financial mutation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/api/v1/plugin/heartbeat": {
      "post": {
        "operationId": "heartbeat",
        "summary": "Heartbeat",
        "parameters": [
          {
            "name": "X-LedgerGuard-Key-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "request_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "server_time": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "request_id",
                    "server_time",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; no source financial mutation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./schemas/woo-heartbeat-1.0.json"
              }
            }
          }
        },
        "security": [
          {
            "WooSignature": []
          }
        ]
      }
    },
...
    "/webhooks/stripe/{destination_id}": {
      "post": {
        "operationId": "stripe_webhook",
        "summary": "Stripe webhook",
        "parameters": [
          {
            "name": "destination_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "receipt_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "receipt_id",
                    "status"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Rejected request; no source financial mutation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "StripeSignature": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "created": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "livemode": {
                    "type": "boolean"
                  },
                  "account": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object"
                  }
                },
                "required": [
                  "id",
                  "type",
                  "created",
                  "livemode",
                  "data"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  }
}